+++ zebra-state/src/service/finalized_state/zebra_db/block/tests/vectors.rs
@@ -1290,6 +1290,7 @@ fn root_at(height: Height) -> BlockCommitmentRoots {
         height,
         sapling_root: sapling::tree::NoteCommitmentTree::default().root(),
         orchard_root: orchard::tree::NoteCommitmentTree::default().root(),
+        auth_data_root: zebra_chain::block::merkle::AuthDataRoot::from([0u8; 32]),
     }
 }